Update PrimitiveStream to use float for Value field
Changed the `InsertRecordInput` struct's `Value` field from `int` to `float64` and updated corresponding code to format the values correctly. This enhances the precision of value data stored and processed by the PrimitiveStream module.

+++ core/types/primitive_stream.go
@@ -2,13 +2,14 @@ package types
 
 import (
 	"context"
+
 	"github.com/golang-sql/civil"
 	"github.com/kwilteam/kwil-db/core/types/transactions"
 )
 
 type InsertRecordInput struct {
 	DateValue civil.Date
-	Value     int
+	Value     float64
 }
 
 type IPrimitiveStream interface {
